+++ CHANGELOG.md
@@ -3,7 +3,9 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
 
 ## [Unreleased]
 ### New
-- Added “maybe” wrappers for `rotate`, `scale` and `translate`.
+- Added “maybe” wrappers for `rotate`, `scale`, `translate`, `union`,
+  `intersection` and `difference`.
+- Added a couple of functions for DFM.
 - Added threaded fasteners and a script showcasing a selection of these.
 - Added `long-hex-diagonal` for computing the long diagonal of a hexagon
   from the short diagonal.